+++ src/Squid.Prism.Engine.Core/Configs/ConfigVariableAttribute.cs
@@ -1,0 +1,7 @@
+namespace Squid.Prism.Engine.Core.Configs;
+
+[AttributeUsage(AttributeTargets.Property)]
+public class ConfigVariableAttribute : Attribute
+{
+    public string? Name { get; set; }
+}