+++ src/Squid.Prism.Server.Core/Attributes/Configs/ConfigVariableAttribute.cs
@@ -1,7 +1,0 @@
-namespace Squid.Prism.Server.Core.Attributes.Configs;
-
-[AttributeUsage(AttributeTargets.Property)]
-public class ConfigVariableAttribute : Attribute
-{
-    public string? Name { get; set; }
-}