+++ src/Squid.Prism.Server.Core/Data/Configs/HttpServerConfig.cs
@@ -1,0 +1,6 @@
+namespace Squid.Prism.Server.Core.Data.Configs;
+
+public class HttpServerConfig
+{
+    public int Port { get; set; } = 8080;
+}