+++ src/Squid.Prism.Server.Core/Types/RouteMethodType.cs
@@ -1,0 +1,12 @@
+namespace Squid.Prism.Server.Core.Types;
+
+public enum RouteMethodType
+{
+    GET,
+    POST,
+    PUT,
+    DELETE,
+    HEAD,
+    OPTIONS,
+    PATCH
+}